# backend/openapi.yaml
openapi: 3.0.3
info:
  title: GeoPingKak API
  description: |
    GeoGuessr resource website API for daily challenges, special maps, and video explanations.

    - Public endpoints require no authentication.
    - Admin endpoints require a `Bearer` token in the `Authorization` header.
    - All endpoints are rate-limited (global 60 req/min, admin 10 req/min).
  version: 1.0.0
servers:
  - url: https://geopingkak-backend-staging-7xgrga7zxa-de.a.run.app
    description: Staging
  - url: https://geopingkak-backend-7xgrga7zxa-de.a.run.app
    description: Production
...
paths:
...
  /api/video-explanations:
    get:
      tags: [Video Explanations]
      summary: Get all video explanations
      description: Returns video explanation data for all dates, including linked challenge URLs.
      operationId: getVideoExplanations
      responses:
        "200":
          description: Array of date-grouped video data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoExplanationDate"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Video Explanations]
      summary: Update video explanations for a date
      description: |
        Partial update for a specific date's video data. Uses Firestore merge mode.
        - Empty fields are ignored (won't overwrite existing data).
        - Automatically looks up and stores `challengeUrl` from daily challenge data.
      operationId: updateVideoExplanations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, maps]
              properties:
                date:
                  type: string
                  format: date
                  description: Target date (`YYYY-MM-DD`)
                maps:
                  type: object
                  description: Map ID to video data mapping
                  additionalProperties:
                    $ref: "#/components/schemas/VideoMapData"
            example:
              date: "2026-03-29"
              maps:
                tw-balanced:
                  livestream: https://youtu.be/example1
                  explanation: https://youtu.be/example2
      responses:
        "200":
          description: Update successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  date:
                    type: string
                  updated_maps:
                    type: array
                    items:
                      type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Challenge URL not found for the given date and map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin token managed via GCP Secret Manager
  schemas:
...
    VideoExplanationDate:
      type: object
      required: [date, maps]
      properties:
        date:
          type: string
          format: date
        maps:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/VideoMapData"
    VideoMapData:
      type: object
      properties:
        livestream:
          type: string
          format: uri
          description: YouTube livestream URL
        explanation:
          type: string
          format: uri
          description: YouTube explanation video URL
        challengeUrl:
          type: string
          format: uri
          description: Auto-populated from daily challenge data
    Error:
      type: object
      required: [error_code, error]
      properties:
        error_code:
          type: string
          description: Machine-readable error code for client consumption
          enum:
            - UNAUTHORIZED
            - INVALID_FORMAT
            - MISSING_FIELD
            - INVALID_FIELD
            - NOT_FOUND
            - RATE_LIMITED
            - METHOD_NOT_ALLOWED
            - INTERNAL_ERROR
        error:
          type: string
          description: HTTP status category (e.g. "Bad Request", "Unauthorized")
        message:
          type: string
          description: Human-readable error detail
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"